FEAT : 구조 변경
