+++ src/main/java/com/hhplus/backend/controller/concert/ConcertController.java
@@ -4,13 +4,13 @@ import com.hhplus.backend.application.concert.ConcertMockFacade;
 import com.hhplus.backend.controller.concert.dto.PayInput;
 import com.hhplus.backend.controller.concert.dto.ReserveSeatInput;
 import com.hhplus.backend.domain.concert.*;
+import com.hhplus.backend.domain.payment.PaymentHistory;
+import com.hhplus.backend.domain.token.UserToken;
 import java.time.LocalDateTime;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
