+++ src/main/java/com/hhplus/backend/domain/payment/PaymentService.java
@@ -1,0 +1,5 @@
+package com.hhplus.backend.domain.payment;
+
+public class PaymentService {
+
+}
