--- src/main/java/com/hhplus/backend/domain/point/PointRepository.java
@@ -1,5 +1,0 @@
-package com.hhplus.backend.domain.point;
-
-public interface PointRepository {
-
-}
