--- src/main/java/com/hhplus/backend/domain/point/PointService.java
@@ -1,4 +1,0 @@
-package com.hhplus.backend.domain.point;
-
-public class PointService {
-}
