+++ src/main/java/com/hhplus/backend/domain/token/TokenRepository.java
@@ -1,0 +1,5 @@
+package com.hhplus.backend.domain.token;
+
+public interface TokenRepository {
+
+}
