+++ src/main/java/com/hhplus/backend/domain/token/TokenService.java
@@ -1,0 +1,5 @@
+package com.hhplus.backend.domain.token;
+
+public class TokenService {
+
+}
