--- src/main/java/com/hhplus/backend/infra/point/PointJpaRepository.java
@@ -1,4 +1,0 @@
-package com.hhplus.backend.infra.point;
-
-public class PointJpaRepository {
-}
