+++ src/main/java/com/hhplus/backend/infra/token/UserTokenJpaRepository.java
@@ -1,0 +1,4 @@
+package com.hhplus.backend.infra.token;
+
+public class UserTokenJpaRepository {
+}
