+++ src/main/java/com/hhplus/backend/infra/token/UserTokenRepositoryImpl.java
@@ -1,0 +1,5 @@
+package com.hhplus.backend.infra.token;
+
+public class UserTokenRepositoryImpl {
+
+}
