+++ src/main/java/com/hhplus/backend/infra/token/entity/UserTokenEntity.java
@@ -1,15 +1,16 @@
-package com.hhplus.backend.infra.point.entity;
+package com.hhplus.backend.infra.token.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
 import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
 
 import java.time.LocalDateTime;
 
 @Entity
 @Data
-@Table(name="POINT_HISTORY")
-public class PointHistoryEntity {
+@Table(name="USER_TOKEN")
+public class UserTokenEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
@@ -19,8 +20,8 @@ public class PointHistoryEntity {
     @Column(name="USER_ID")
     private Long userId;
 
-    @Column(name="POINT")
-    private int point;
+    @Column(name="ACCESS_TOKEN")
+    private String accessToken;
 
     @Column(name="STATUS")
     private String status;
@@ -29,4 +30,7 @@ public class PointHistoryEntity {
     @Column(name="CREATED_AT")
     private LocalDateTime createdAt;
 
+    @LastModifiedDate
+    @Column(name="UPDATED_AT")
+    private LocalDateTime updatedAt;
 }
