+++ src/main/java/com/hhplus/backend/infra/user/entity/PointHistoryEntity.java
@@ -1,16 +1,15 @@
-package com.hhplus.backend.infra.concert.entity;
+package com.hhplus.backend.infra.user.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 
 import java.time.LocalDateTime;
 
 @Entity
 @Data
-@Table(name="USER_TOKEN")
-public class UserTokenEntity {
+@Table(name="POINT_HISTORY")
+public class PointHistoryEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
@@ -20,8 +19,8 @@ public class UserTokenEntity {
     @Column(name="USER_ID")
     private Long userId;
 
-    @Column(name="ACCESS_TOKEN")
-    private String accessToken;
+    @Column(name="POINT")
+    private int point;
 
     @Column(name="STATUS")
     private String status;
@@ -30,7 +29,4 @@ public class UserTokenEntity {
     @Column(name="CREATED_AT")
     private LocalDateTime createdAt;
 
-    @LastModifiedDate
-    @Column(name="UPDATED_AT")
-    private LocalDateTime updatedAt;
 }
